# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Exam Backend API",
    "description": "API documentation for Sindre Gangeskar's Exam Project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "adminAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Include **Bearer** prefix in the value field with token - e.g **Bearer abc123**"
    },
    "userAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Include **Bearer** prefix in the value field with token - e.g **Bearer abc123**"
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "description": "A guest can log in to a **registered** account with **Username** or **Email**",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User credentials for logging in. The API accepts either the **username** or **email**, and allows the user to choose which to use as it doesn't require both to be in the body.",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "johndoe"
                },
                "password": {
                  "type": "string",
                  "example": "secret"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@email.com"
                }
              },
              "required": [
                "username",
                "password"
              ]
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the user successully logs in",
            "schema": {
              "$ref": "#/definitions/authPostLoginSuccess"
            }
          },
          "400": {
            "description": "This response is returned when the user has provided incorrect or invalid values or properties in the body",
            "schema": {
              "$ref": "#/definitions/authPostLoginUserError"
            }
          },
          "404": {
            "description": "This response is returned when the API fails to find user with provided **username** or **email**",
            "schema": {
              "$ref": "#/definitions/authPostLoginNotFound"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** while the API is trying to find user data",
            "schema": {
              "$ref": "#/definitions/authPostLoginInternalError"
            }
          }
        }
      }
    },
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "An **admin** can retrieve all existing users",
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully finds all users regsitered. If no users are found, an **empty array** will be returned",
            "schema": {
              "$ref": "#/definitions/usersGetSuccess"
            }
          },
          "401": {
            "description": "This response is returned when the user lacks **admin** privileges or isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/adminUnauthorized"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to get all users",
            "schema": {
              "$ref": "#/definitions/usersGetInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          }
        ]
      }
    },
    "/users/user/details": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "A logged in **user** can retrieve its own details",
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully finds user details for the logged in user",
            "schema": {
              "$ref": "#/definitions/userGetByIdSuccess"
            }
          },
          "401": {
            "description": "This response is returned when the user isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/unauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API fails to find user details for the logged in user",
            "schema": {
              "$ref": "#/definitions/userGetByIdNotFound"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to find user details for the logged in user",
            "schema": {
              "$ref": "#/definitions/userGetByIdInternalError"
            }
          }
        },
        "security": [
          {
            "userAuth": [],
            "adminAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "An **admin** can retrieve a user\\'s details by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The user id"
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully finds the user with the provided id",
            "schema": {
              "$ref": "#/definitions/userGetByIdSuccess"
            }
          },
          "401": {
            "description": "This response is returned when the user lacks **admin** privileges or isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/adminUnauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API fails to find a user with the provided id",
            "schema": {
              "$ref": "#/definitions/userGetByIdNotFound"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to find a user with the provided id",
            "schema": {
              "$ref": "#/definitions/userGetByIdInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "An **admin** can update a user\\'s data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The user id"
          },
          {
            "name": "body",
            "in": "body",
            "description": "The available properties for updating a user's details",
            "schema": {
              "type": "object",
              "properties": {
                "RoleId": {
                  "type": "number",
                  "example": 1
                },
                "email": {
                  "type": "string",
                  "example": "new@email.com"
                },
                "firstname": {
                  "type": "string",
                  "example": "John"
                },
                "lastname": {
                  "type": "string",
                  "example": "Smith"
                },
                "address": {
                  "type": "string",
                  "example": "New address"
                },
                "phone": {
                  "type": "string",
                  "example": "123456789"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully updates the user data",
            "schema": {
              "$ref": "#/definitions/userUpdateByIdSuccess"
            }
          },
          "400": {
            "description": "This response is returned when the **admin** user provides incorrect or invalid values or properties in the body",
            "schema": {
              "$ref": "#/definitions/userUpdateByIdUserError"
            }
          },
          "401": {
            "description": "This response is returned when the user lacks **admin** privileges or isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/adminUnauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API fails to find a user with the provided user id, or role id",
            "schema": {
              "$ref": "#/definitions/userUpdateByIdNotFound"
            }
          },
          "409": {
            "description": "This response is returned when the user provides an **email** that is already in use by another user",
            "schema": {
              "$ref": "#/definitions/userUpdateByIdConflict"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to find a user with the provided id",
            "schema": {
              "$ref": "#/definitions/userUpdateByIdInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "unauthorized": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 401
            },
            "result": {
              "type": "array",
              "example": [
                "Login required",
                "Invalid token, please log in again",
                "Session has expired, please log in again"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidJWTError",
                "InvalidCredentialsError",
                "UnauthorizedError",
                "TokenExpiredError"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "adminUnauthorized": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 401
            },
            "result": {
              "type": "array",
              "example": [
                "Login required",
                "Admin privileges required",
                "Invalid token, please log in again",
                "Session has expired, please log in again"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidJWTError",
                "InvalidCredentialsError",
                "UnauthorizedError",
                "TokenExpiredError"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "usersGetSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "array",
              "example": [
                "Successfully found users",
                "No users found"
              ],
              "items": {
                "type": "string"
              }
            },
            "users": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "username": {
                    "type": "string",
                    "example": "user"
                  },
                  "firstname": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastname": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@email.com"
                  },
                  "address": {
                    "type": "string",
                    "example": "Example Address"
                  },
                  "phone": {
                    "type": "string",
                    "example": "123456789"
                  },
                  "totalPurchases": {
                    "type": "number",
                    "example": 2
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2024-11-05T13:53:25.000Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2024-11-05T13:53:25.000Z"
                  },
                  "RoleId": {
                    "type": "number",
                    "example": 2
                  },
                  "role": {
                    "type": "string",
                    "example": "User"
                  },
                  "MembershipId": {
                    "type": "number",
                    "example": 1
                  },
                  "membership": {
                    "type": "string",
                    "example": "Bronze"
                  }
                }
              }
            }
          }
        }
      }
    },
    "usersGetInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "message": {
              "type": "string",
              "example": "An internal server error has occurred while trying to get all users"
            }
          }
        }
      }
    },
...
    "authPostLoginSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully logged in"
            },
            "token": {
              "type": "string",
              "example": "<jwt>"
            },
            "user": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "john smith"
                },
                "role": {
                  "type": "string",
                  "example": "user"
                },
                "email": {
                  "type": "string",
                  "example": "johnsmith@email.com"
                }
              }
            }
          }
        }
      }
    },
    "authPostLoginUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 401
            },
            "result": {
              "type": "array",
              "example": [
                "Required properties are missing",
                "username or email is required and must be a non-empty string",
                "password is required and must be a non-empty string",
                "Invalid credentials, please try again"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidCredentialsError",
                "MissingCredentialsError",
                "InvalidStringError"
              ],
              "items": {
                "type": "string"
              }
            },
            "missingInBody": {
              "type": "array",
              "example": [
                "username",
                "password",
                "email"
              ],
              "items": {
                "type": "string"
              }
            },
            "invalidProperties": {
              "type": "array",
              "example": [
                "ExampleInvalidProperty1",
                "ExampleInvalidProperty2"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "authPostLoginNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "string",
              "example": "No user with provided email or username exists"
            },
            "error": {
              "type": "string",
              "example": "UserNotFoundError"
            }
          }
        }
      }
    },
    "authPostLoginInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "message": {
              "type": "string",
              "example": "An internal server error has occurred while trying to find user data"
            }
          }
        }
      }
    },
    "userGetByIdSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully found user"
            },
            "user": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "username": {
                  "type": "string",
                  "example": "johnSmith"
                },
                "firstname": {
                  "type": "string",
                  "example": "John"
                },
                "lastname": {
                  "type": "string",
                  "example": "Smith"
                },
                "email": {
                  "type": "string",
                  "example": "johnsmith@email.com"
                },
                "address": {
                  "type": "string",
                  "example": "Example Address"
                },
                "phone": {
                  "type": "string",
                  "example": "123456789"
                },
                "totalPurchases": {
                  "type": "number",
                  "example": 3
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-12-05T13:20:59.000Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2024-12-05T13:20:59.000Z"
                },
                "RoleId": {
                  "type": "number",
                  "example": 2
                },
                "role": {
                  "type": "string",
                  "example": "User"
                },
                "MembershipId": {
                  "type": "number",
                  "example": 1
                },
                "membership": {
                  "type": "string",
                  "example": "Bronze"
                }
              }
            }
          }
        }
      }
    },
    "userGetByIdNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "string",
              "example": "No user with provided id exists"
            },
            "error": {
              "type": "string",
              "example": "UserNotFoundError"
            }
          }
        }
      }
    },
    "userGetByIdInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to get user with provided id"
            }
          }
        }
      }
    },
    "userUpdateByIdSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully updated user"
            }
          }
        }
      }
    },
    "userUpdateByIdUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 400
            },
            "result": {
              "type": "array",
              "example": [
                "Please remove invalid properties",
                "At least one property must be present",
                "firstname must be a non-empty string",
                "lastname must be a non-empty string",
                "address must be a non-empty string",
                "phone must be a non-empty string with numbers only (spaces allowed)",
                "email must be a non-empty string which follows the correct email format, e.g example@email.com",
                "RoleId must be a number"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidPropertyError",
                "BodyEmptyUserError",
                "InvalidStringError",
                "InvalidEmailError",
                "InvalidPhoneError"
              ],
              "items": {
                "type": "string"
              }
            },
            "validProperties": {
              "type": "array",
              "example": [
                "RoleId",
                "firstname",
                "lastname",
                "email",
                "phone"
              ],
              "items": {
                "type": "string"
              }
            },
            "invalidProperties": {
              "type": "array",
              "example": [
                "InvalidPropertyExample1, InvalidPropertyExample2"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "userUpdateByIdNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "array",
              "example": [
                "No user with provided id exists",
                "No role with provided id exists"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "UserNotFoundError",
                "RoleNotFoundError"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "userUpdateByIdInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "message": {
              "type": "string",
              "example": "An internal server error has occurred while trying to update user data"
            }
          }
        }
      }
    },
...
  }
}